fix run.sh to add . to PYTHONPATH
Does anyone still use this old script besides me?  It is kinda yucky and I
think it needs this change to work with the new virtualenv hotness.

Also: what do ya'll think about possibly switching to Poetry for our
python builds?  It is really nice (and apparently the new cool way to
make projects for pypi) and hides the virtualenv goo.  If you are
interested I'm willing to send in a PR for your consideration.  Would
make the env for devs much easier I think - also nicely packages for
pypi and removes the need for requirements.txt.